voice-chat-be/requirements.txt
# FastAPI & Web Server
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
aiofiles
python-multipart
websockets

# Database
SQLAlchemy
pymysql
pymilvus

# AI & ML Core - Gemma 호환성을 위한 최신 안정 버전 조합
torch==2.2.1
transformers==4.40.1
sentence-transformers==2.7.0
accelerate==0.29.3
protobuf==3.20.3
sentencepiece==0.2.0
huggingface-hub

# STT (Speech-to-Text)
openai-whisper

# Audio Processing
librosa
soundfile
av

# Utilities
numpy
pandas
scipy
scikit-learn
tqdm
loguru
requests
certifi
charset-normalizer
idna
urllib3
yarl
aiohttp
aiosignal
async-timeout
attrs
frozenlist
multidict
six
boto3
botocore
s3transfer
jmespath
python-dateutil
pytz
tzdata

# TTS Dependencies (based on MeloTTS/OpenVoice)
unidic-lite
jamo
pykakasi
g2pkk
eng_to_ipa
inflect
ko_pron
protobuf
einops
omegaconf
prettytable
pytorch-lightning
PyYAML
phonemizer
alias-free-torch
audioread
gradio
gradio_client
altair
Markdown
ffmpy
Jinja2
MarkupSafe
Pillow
pygments
soupsieve
beautifulsoup4
orjson
itsdangerous
werkzeug
weasel
coloredlogs
emoji
num2words
langcodes
regex
psutil
filelock
sniffio
httpcore
httpx
h11
anyio
exceptiongroup
starlette
typing_extensions
confection
cymem
murmurhash
preshed
spacy
spacy-legacy
spacy-loggers
srsly
thinc
wasabi
catalogue
blinker
click
colorama
decorator
Deprecated
distro
fsspec
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
importlib_resources
jiter
jsonschema
jsonschema-specifications
lazy_loader
llvmlite
marisa-trie
markdown-it-py
mdurl
more-itertools
mpmath
msgpack
narwhals
networkx
nltk
numba
onnxruntime
packaging
plac
platformdirs
pooch
propcache
proto-plus
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydub
pygame
pynndescent
pyparsing
pypinyin
pyreadline3
pysbd
python-crfsuite
python-mecab-ko
python-mecab-ko-dic
referencing
resampy
rich
rpds-py
rsa
ruff
safehttpx
safetensors
semantic-version
shellingham
smart_open
soxr
SudachiDict-core
SudachiPy
sympy
tensorboard
tensorboard-data-server
threadpoolctl
tiktoken
tokenizers
tomlkit
torchaudio
torchvision
trainer
typeguard
typer
typing-inspection
tzlocal
ujson
umap-learn
Unidecode
unidic
wavmark
whisper-timestamped
win32_setctime
wrapt
zipp
